+++ src/main/java/Heuristic/PopulationBasedMetaHeuristic/GeneticAlgorithm/SearchMethod.java
@@ -1,5 +1,5 @@
 package Heuristic.PopulationBasedMetaHeuristic.GeneticAlgorithm;
 
 public interface SearchMethod {
-    double run();
+    int run();
 }
