+++ src/main/java/Heuristic/PopulationBasedMetaHeuristic/SetOfMethods/Replacement/Replacement.java
@@ -26,17 +26,17 @@ public class Replacement {
         int[] aiOffpsringMemoryIndices = IntStream.range(iPopulationSize, iTotalPopulationSize).toArray();
 
         // elitism replacing worst offspring with best solution if not in offspring
-        double[] adTotalPopulationCosts = new double[iTotalPopulationSize];
+        int[] adTotalPopulationCosts = new int[iTotalPopulationSize];
 
-        double dBestSolutionCost = Integer.MIN_VALUE;
-        double dWorstOffspringCost =Integer.MAX_VALUE;
+        int dBestSolutionCost = Integer.MIN_VALUE;
+        int dWorstOffspringCost =Integer.MAX_VALUE;
         int bestIndex = -1;
         int worstOffspringIndex = -1;
 
         // evaluate the objective function value (cost) of each solution from both parent and offspring populations
         for(int iMemoryIndex = 0; iMemoryIndex < iTotalPopulationSize; iMemoryIndex++) {
 
-            double dSolutionCost = oProblem.getObjectiveFunctionValue(iMemoryIndex);
+            int dSolutionCost = oProblem.getObjectiveFunctionValue(iMemoryIndex);
             adTotalPopulationCosts[iMemoryIndex] = dSolutionCost;
 
             // update index of best solution, favouring offspring solutions
