+++ src/main/java/Heuristic/SinglePointMetaHeuristic/HillClimb/FirstImprovementHC.java
@@ -7,10 +7,10 @@ public class FirstImprovementHC extends HeuristicMethods {
 
     @Override
     public void applyHeuristic(Problem problem) {
-        double bestEval=problem.getObjectiveFunctionValue(CURRENT_SOLUTION_INDEX);
+        int bestEval=problem.getObjectiveFunctionValue(CURRENT_SOLUTION_INDEX);
         for(int j=0;j<problem.getNumberOfVariables(CURRENT_SOLUTION_INDEX);j++){
             problem.bitFlip(CURRENT_SOLUTION_INDEX,j);
-            double tmpEval= problem.getObjectiveFunctionValue(CURRENT_SOLUTION_INDEX);
+            int tmpEval= problem.getObjectiveFunctionValue(CURRENT_SOLUTION_INDEX);
             if(tmpEval>bestEval)
                 bestEval=tmpEval;
             else
